Initial commit - sample of defensive serializer
A sample application with a DefensiveMappingPdxSerializer.

This serializer extends Spring Data GemFires MappingPdxSerializer. It adds the
capability to detect a cycle in the object being serialized and throw an
exception if a cycle is detected. This replaces a StackOverFlowError with a
more informative IllegalStateException.

This serializer probably comes with a small peformance penalty, because it
updates a set for every object that is serialized.